Add GameCube support
