+++ source/application_gc.cpp
@@ -1,0 +1,14 @@
+#include "application_gc.h"
+
+/**
+ * Constructor for the ApplicationGc class.
+ */
+ApplicationGc::ApplicationGc() : Application() {
+}
+
+/**
+ * Destructor for the ApplicationGc class.
+ */
+ApplicationGc::~ApplicationGc()
+{
+}
