+++ source/main.cpp
@@ -1,5 +1,11 @@
 #include <memory>
-#include "application.h"
+#if defined(HW_RVL)
+    #include "application_wii.h"
+    #define APPLICATIONCLASS ApplicationWii
+#elif defined(HW_DOL)
+    #include "application_gc.h"
+    #define APPLICATIONCLASS ApplicationGc
+#endif
 
 /**
  * Entry point.
@@ -7,7 +13,7 @@
  * @param argv An array of null-terminated strings representing command-line arguments.
  */
 int main(int argc, char *argv[]) {
-    auto app = std::make_unique<Application>();
+    auto app = std::make_unique<APPLICATIONCLASS>();
     if(argc > 0 && argv[0] != nullptr) {
         app->SetPath(argv[0]);
     }
